refactor(react): refactored utils into util dir
