+++ packages/botonic-react/src/components/message-template.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { renderComponent } from '../utils'
+import { renderComponent } from '../util/react'
 
 export const MessageTemplate = props => {
   const renderBrowser = () => null
