+++ packages/botonic-react/src/components/pic.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import styled from 'styled-components'
 import { COLORS, WEBCHAT } from '../constants'
-import { renderComponent } from '../utils'
+import { renderComponent } from '../util/react'
 
 const PicStyled = styled.div`
   border-top-left-radius: 8px;
