+++ packages/botonic-react/src/index.js
@@ -7,4 +7,4 @@ export { WebviewApp } from './webview'
 export * from './components'
 export { BotonicInputTester, BotonicOutputTester } from './botonic-tester'
 export { msgToBotonic, msgsToBotonic } from './msg-to-botonic'
-export { staticAsset } from './utils'
+export { staticAsset } from './util/environment'
