+++ packages/botonic-react/src/webchat/components/common.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import { resolveImage } from '../../utils'
+
+import { resolveImage } from '../../util/environment'
 
 const PointerImage = styled.img`
   cursor: pointer;
